Ensure directory exists before deleting it
git-svn-id: https://svn.eiffel.com/eiffelstudio-public/trunk@98681 8089f293-4706-0410-a29e-feb5c42a2edf

+++ Src/examples/vision2/widget_test/retrieve_flat_short.bat
@@ -1,4 +1,4 @@
-rd /Q /S %TEMP%\vision2_precompile
+if exist %TEMP%\vision2_precompile rd /Q /S %TEMP%\vision2_precompile
 mkdir %TEMP%\vision2_precompile
 %ISE_EIFFEL%\studio\spec\%ISE_PLATFORM%\bin\ecb -config vision2.ecf -local -project_path %TEMP%\vision2_precompile
